:tractor: Split screens into individual files
